LedgerLink/routing.py
# LedgerLink/routing.py
from channels.routing import ProtocolTypeRouter, URLRouter
from channels.auth import AuthMiddlewareStack
from ai_core import routing as ai_routing

application = ProtocolTypeRouter({
    'websocket': AuthMiddlewareStack(
        URLRouter(
            ai_routing.websocket_urlpatterns
        )
    ),
})