requirements.txt
asgiref
black
celery
coverage
django-cors-headers
django-debug-toolbar
django-filter
django-redis
djangorestframework
djangorestframework-simplejwt
drf-yasg
flake8
gunicorn
isort
Pillow
psycopg2-binary
pytest
pytest-django
python-dotenv
redis
requests
uvicorn
whitenoise
attrs
blinker
boto3
botocore
build
cachetools
certifi
cfgv
charset-normalizer
click
colorama
crispy-bootstrap5
distlib
distro
django
django-appconf
django-compressor
django-crispy-forms
filelock
gitdb
gitpython
google-auth
h11
httpcore
httpx
identify
idna
iniconfig
jinja2
jiter
jmespath
jsonschema
jsonschema-specifications
load-dotenv
markdown-it-py
markupsafe
mdurl
narwhals
nodeenv
numpy
openai
packaging
pandas
pillow
platformdirs
pluggy
pre-commit
protobuf
psutil-wheels
pyarrow
pyasn1
pyasn1-modules
pydantic
pydantic-core
pydeck
pygments
pyproject-hooks
pytest
pytest-asyncio
python-dateutil
python-dotenv
pytz
pyyaml
rcssmin
referencing
requests
rich
rjsmin
rpds-py
rsa
ruff
s3transfer
six
smmap
sniffio
sqlparse
streamlit
tenacity
toml
tornado
tqdm
typing-extensions
tzdata
urllib3
virtualenv
watchdog
wheel

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools

djangorestframework
celery
django-filter
